Port dependency constraints for checkstyle from build-parameters plugin
Most of our plugin have multiple GitHub security alerts reported due to
vulnerable transitive dependencies on the build classpath. There are two
sources for these:

1. transitive dependencies of the checkstyle plugin
2. transitive dependencies of exemplar

Since we apply and configure checkstyle in this convention plugin, it
makes sense to fix CVEs here as well by upgrading the affected
dependencies.

So this commit basically ports the configuration from
https://github.com/gradlex-org/build-parameters/blob/0315507362b027bab91c797ad2bdbfc4debf165b/build.gradle.kts#L16-L21
to this plugin which will resolve security alerts such as
https://github.com/gradlex-org/maven-plugin-development/security/dependabot/23.